backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
pydantic>=2.10.0
pydantic-settings>=2.7.0
sentence-transformers>=3.4.0
chromadb>=0.6.0
PyPDF2>=3.0.0
python-docx>=1.1.0
httpx>=0.28.0
whisperx>=3.3.4
